+++ eds-aic-20230930/core/Models/CreateScheduledTaskRequest.cs
@@ -72,6 +72,14 @@ namespace AlibabaCloud.SDK.Eds_aic20230930.Models
             public int? MaxSteps { get; set; }
 
             /// <summary>
+            /// <b>Example:</b>
+            /// <para>[&quot;sk-abc&quot;]</para>
+            /// </summary>
+            [NameInMap("Skills")]
+            [Validation(Required=false)]
+            public List<string> Skills { get; set; }
+
+            /// <summary>
             /// <para>The timeout period, in seconds.</para>
             /// 
             /// <b>Example:</b>